Fix the requested changes
